little changes in some method parameter types

+++ push_swap.h
@@ -2,7 +2,6 @@
 # define PUSH_SWAP_H
 
 # include <limits.h>
-# include <stddef.h>
 # include <stdio.h>
 # include <stdlib.h>
 
@@ -22,14 +21,14 @@ void		print_int_arr(int *arr, int size);
 int			*validate_input(char **args, int size);
 
 // stack operations part
-t_stack		*create_stack(int *arr, size_t size, size_t capacity);
+t_stack		*create_stack(int *arr, int size, int capacity);
 void		swap(t_stack *stack);
 void		rotate(t_stack *stack);
 void		rrotate(t_stack *stack);
 void		push(t_stack *a, t_stack *b);
 
 // I think main sorting logic will go here (might rename file and method later)
-void		stack_controller(int *arr_a, int *arr_b, size_t size);
+void		stack_controller(int *arr_a, int *arr_b, int size);
 
 // method for testing if operations work properly
 void		stack_operations_tests(t_stack *stack_a, t_stack *stack_b);
